// src/common/Types/CommonTypes.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
import { Rule } from 'antd/es/form';
import { ValidateStatus } from 'antd/es/form/FormItem';
import { NamePath } from 'antd/es/form/interface';
import { UploadFile, UploadListType } from 'antd/es/upload/interface';
import { Dispatch, SetStateAction } from 'react';

export interface IAccountHead {
  id: number;
  code: string;
  parent_id: number;
  name: string;
  group_name: string;
}
export type commonProps = {
  handleSelect?: (id: number) => any;
  onSelectEmployee?: any;
  onCategorySelect?: any;
  id?: any;
  accountSelectID?: number;
  setAccountDetails?: Dispatch<SetStateAction<any[]>>;
  handleClient?: any;
  handleStaff?: any;
  staffId?: any;
  selectedValue?: any;
  supplierId?: any;
  productId?: any;
  showAll?: boolean;
  name?: NamePath;
  label?: string;
  disabled?: boolean;
  initialValue?: string;
  defaultValue?: number | string;
  placeholder?: string;
  required?: boolean;
  size?: number;
  smSize?: number;
  xlSize?: number;
  mdSize?: number;
  xsSize?: number;
  small?: boolean;
  rules?: Rule[];
  dependencies?: NamePath[];
  mode?: 'multiple' | 'tags' | undefined;
  loading?: boolean;
  multiple?: boolean;
  value?: string;
  onChange?: (e: any) => any;
  handleMemberChange?: (value?: any) => void;
  format?: any;
  picker?: 'time' | 'date' | 'year' | 'week' | 'month' | 'quarter' | undefined;
  onSearch?: (value: string) => void;
  maxCount?: number;
  namePrefix?: string;
  help?: string;
  prefix?: string;
  helperText?: string;
  hasFeedback?: boolean;
  data?: any;
  handleQuantity?: ((value?: any) => void) | undefined;
  handleAttribute?: (value?: any) => void;
  quantity?: number;
  accept?: string;
  picture?: UploadListType;
  fileList?: UploadFile[];
  setFileList?: Dispatch<SetStateAction<UploadFile[]>>;
  style?: React.CSSProperties;
  rows?: number | undefined;
  validateStatus?: ValidateStatus;
  onBlur?: any;
  onSelect?: any;
  rangePicker?: boolean;
  year?: boolean;
};